Created loadFromFile()
Added a public loadFromFile method to Environment so data can be pulled
in from an external config file.
Along with the main method there are four private methods to draw data
out of a given line in the config; readSizeLine(), readColourLine(),
readObjectLine() and readEntityLine(). These methods take an iss formed
prior and extract the appropriate data to set a member value or create
new Objects or Entities.