Implemented the PathTo() and Explore() methods
I've added a queue of position vectors to Colonist that will be used to
store the route to a destination.
I altered the Colonist delcaration so radius and speed are set within
the constructor for uniformity.
I had to add 'srand()' to main() so I could generate a seed for rand()
functionality.
The PathTo() method has been introduced that moves the Colonist
direction to an input position.
The Explore() method is currently set up to randomly add a position in
front of the Colonist within a 75° cone to the m_path queue.
The update loop has been set to permenantly stay in an EXPLORE state and
wander aimlessly around the Environment.